requirements.txt
Flask
flask-cors
flask-sqlalchemy
pymysql
python-dotenv
